requirements.txt
pylibjpeg[all]
numpy
